requirements.txt
numpy==1.21.1
ipython==7.21.0
addict==2.4.0
future==0.18.2
lmdb==1.3.0
opencv-python==4.5.5.64
Pillow==9.1.0
pyyaml==6.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
scipy==1.8.0
scikit-image==0.19.2
matplotlib==3.5.1
